requirements.txt
flask
langchain
langgraph
google-generativeai
langchain-google-genai
python-dotenv